+++ navigator/r2-navigator-swift/EPUB/EPUBSpreadView.swift
@@ -93,16 +93,7 @@ class EPUBSpreadView: UIView, Loggable, PageView {
         registerJSMessages()
 
         NotificationCenter.default.addObserver(self, selector: #selector(voiceOverStatusDidChange), name: Notification.Name(UIAccessibilityVoiceOverStatusChanged), object: nil)
-        
-        if editingActions.canPerformAction(EditingAction.share.selector) {
-            UIMenuController.shared.menuItems = [
-                UIMenuItem(
-                    title: R2NavigatorLocalizedString("EditingAction.share"),
-                    action: #selector(shareSelection)
-                )
-            ]
-        }
-        
+
         updateActivityIndicator()
         loadSpread()
     }